+++ modules/segmentation_module/segmentation.py
@@ -1,9 +1,9 @@
 import logging
 import os
 import concurrent
-from PyQt5.QtWidgets import QFileDialog, QCheckBox, QPushButton, QVBoxLayout, QHBoxLayout, QWidget, QGroupBox, QRadioButton, QApplication, QSpinBox, QFormLayout, QLabel, QLineEdit, QComboBox
+from PyQt5.QtWidgets import QCheckBox, QPushButton, QVBoxLayout, QHBoxLayout, QWidget, QGroupBox, QRadioButton, QApplication, QSpinBox, QFormLayout, QLabel, QLineEdit, QComboBox
 from PyQt5.QtCore import Qt, QRegExp
-from PyQt5.QtGui import QCursor, QIntValidator, QRegExpValidator
+from PyQt5.QtGui import QCursor, QDoubleValidator, QRegExpValidator
 from modules.segmentation_module import segmentation_functions as f
 from general import general_functions as gf
 import torch
@@ -18,15 +18,18 @@ class Segmentation(QWidget):
 
         self.pipeline_layout = pipeline_layout
 
-        label_documentation = gf.CollapsibleLabel('',collapsed=True)
+        label_documentation = gf.CollapsibleLabel('',  collapsed=True)
         label_documentation.setText('For each input image,  perform cell segmentation using <a href="https://www.cellpose.org/">cellpose</a> and save the resulting mask.<br>' +
                                     'Input images must have X and Y axes and can optionally have C, Z and/or T axes (Z axis will be projected and only the chosen channel will be selected before performing segmentation).<br>' +
-                                    'A "cellpose model" can be obtained by finetuning a pretrained cellpose model on a collection of annotated images similar to the input images (see section "Training" in cellpose documentation <a href="https://cellpose.readthedocs.io">https://cellpose.readthedocs.io</a>).')
+                                    '<h3>Cellpose model types</h3>' +
+                                    '<b>Built-in models</b><br>' +
+                                    'Built-in models cyto, cyto2, nuclei, tissuenet and livecell are available. These models require a parameter "diameter", which should correspond to the expected diameter of the objects to segment. With cyto, cyto2 or nuclei model types, the diameter can be estimated using cellpose built-in model by setting the parameter "diameter" to 0. For more information, see section "Models" in cellpose documentation <a href="https://cellpose.readthedocs.io">https://cellpose.readthedocs.io</a>' +
+                                    '<br>' +
+                                    '<b>User trained model</b><br>' +
+                                    'A user trained model can be obtained by finetuning a pretrained cellpose model on a collection of annotated images similar to the input images (see section "Training" in cellpose documentation <a href="https://cellpose.readthedocs.io">https://cellpose.readthedocs.io</a>). When using a user trained model it is not possible to choose the diameter, which is set to the median diameter estimated on the training set.')
 
         self.image_list = gf.FileListWidget(filetypes=gf.imagetypes, filenames_filter='_BF')
         self.image_list.file_list_changed.connect(self.image_list_changed)
-
-        self.selected_model = gf.FileLineEdit()
 
         self.use_input_folder = QRadioButton("Use input image folder")
         self.use_input_folder.setChecked(True)
@@ -47,10 +50,37 @@ class Segmentation(QWidget):
         self.output_filename_label.setEnabled(False)
         self.output_filename_label.textChanged.connect(self.output_filename_label.setToolTip)
 
+        self.cellpose_model_type = QComboBox()
+        self.cellpose_model_type.addItem("User trained model")
+        self.cellpose_model_type.addItem("cyto")
+        self.cellpose_model_type.addItem("cyto2")
+        self.cellpose_model_type.addItem("nuclei")
+        self.cellpose_model_type.addItem("tissuenet")
+        self.cellpose_model_type.addItem("livecell")
+        self.cellpose_model_type.setCurrentText("User trained model")
+        self.cellpose_model_type.currentTextChanged.connect(self.cellpose_model_type_changed)
+        self.cellpose_user_model = gf.FileLineEdit()
+        self.cellpose_user_model_label = QLabel("Model:")
+        self.cellpose_diameter = QSpinBox()
+        self.cellpose_diameter.setMinimum(0)
+        self.cellpose_diameter.setMaximum(1000)
+        self.cellpose_diameter.setValue(0)
+        self.cellpose_diameter.setToolTip('Expected cell diameter (pixel). If 0, use cellpose built-in model to estimate diameter (available only for cyto, cyto2 and nuclei models).')
+        self.cellpose_diameter.setVisible(False)
+        self.cellpose_diameter_label = QLabel("Diameter:")
+        self.cellpose_diameter_label.setVisible(False)
+        self.cellpose_cellprob_threshold = QLineEdit(placeholderText='0.0')
+        self.cellpose_cellprob_threshold.setValidator(QDoubleValidator(decimals=2))
+        self.cellpose_cellprob_threshold.validator().setNotation(QDoubleValidator.StandardNotation)
+        self.cellpose_flow_threshold = QLineEdit(placeholderText='0.4')
+        self.cellpose_flow_threshold.setValidator(QDoubleValidator(decimals=2))
+        self.cellpose_flow_threshold.validator().setNotation(QDoubleValidator.StandardNotation)
+
         self.channel_position = QSpinBox()
         self.channel_position.setMinimum(0)
         self.channel_position.setMaximum(100)
         self.channel_position.setValue(0)
+
         # Z-Projection range
         # only bestZ
         self.projection_mode_bestZ = QRadioButton("Z section with best focus")
@@ -129,12 +159,6 @@ class Segmentation(QWidget):
             groupbox.setLayout(layout2)
             layout.addWidget(groupbox)
 
-        groupbox = QGroupBox("Cellpose model")
-        layout2 = QVBoxLayout()
-        layout2.addWidget(self.selected_model)
-        groupbox.setLayout(layout2)
-        layout.addWidget(groupbox)
-
         groupbox = QGroupBox("Output")
         layout2 = QVBoxLayout()
         if not self.pipeline_layout:
@@ -159,14 +183,27 @@ class Segmentation(QWidget):
         layout.addWidget(groupbox)
 
         groupbox = QGroupBox("Options")
-        layout3 = QFormLayout()
-        layout3.setLabelAlignment(Qt.AlignLeft)
-        layout3.setFormAlignment(Qt.AlignLeft)
+        layout3 = QVBoxLayout()
+
+        groupbox2 = QGroupBox("Segmentation method")
+        layout4 = QFormLayout()
+        layout4.addRow("Model type:", self.cellpose_model_type)
+        layout4.addRow(self.cellpose_user_model_label, self.cellpose_user_model)
+        layout4.addRow(self.cellpose_diameter_label, self.cellpose_diameter)
+        collapsible = gf.CollapsibleWidget("", collapsed_icon="▶ [show more]", expanded_icon="▼ [hide]", expanded=False)
+        layout5 = QFormLayout()
+        collapsible.content.setLayout(layout5)
+        layout5.addRow("Cellprob threshold:", self.cellpose_cellprob_threshold)
+        layout5.addRow("Flow threshold:", self.cellpose_flow_threshold)
+        layout4.addRow(collapsible)
+        groupbox2.setLayout(layout4)
+        layout3.addWidget(groupbox2)
+
         groupbox2 = QGroupBox("If multiple channels:")
         layout4 = QFormLayout()
         layout4.addRow("Channel position:", self.channel_position)
         groupbox2.setLayout(layout4)
-        layout3.addRow(groupbox2)
+        layout3.addWidget(groupbox2)
 
         groupbox2 = QGroupBox("If multiple z:")
         layout4 = QFormLayout()
@@ -202,7 +239,7 @@ class Segmentation(QWidget):
         layout4.addRow("Projection range:", widget)
         layout4.addRow("Projection type:", self.projection_type)
         groupbox2.setLayout(layout4)
-        layout3.addRow(groupbox2)
+        layout3.addWidget(groupbox2)
         groupbox.setLayout(layout3)
         layout.addWidget(groupbox)
 
@@ -224,6 +261,18 @@ class Segmentation(QWidget):
         self.logger = logging.getLogger(__name__)
 
         self.update_output_filename_label()
+
+    def cellpose_model_type_changed(self, model):
+        if model == 'User trained model':
+            self.cellpose_user_model_label.setVisible(True)
+            self.cellpose_user_model.setVisible(True)
+            self.cellpose_diameter_label.setVisible(False)
+            self.cellpose_diameter.setVisible(False)
+        else:
+            self.cellpose_user_model_label.setVisible(False)
+            self.cellpose_user_model.setVisible(False)
+            self.cellpose_diameter_label.setVisible(True)
+            self.cellpose_diameter.setVisible(True)
 
     def image_list_changed(self):
         if self.image_list.count() > 1:
@@ -260,7 +309,11 @@ class Segmentation(QWidget):
             'use_input_folder': self.use_input_folder.isChecked(),
             'use_custom_folder': self.use_custom_folder.isChecked(),
             'output_folder': self.output_folder.text(),
-            'selected_model': self.selected_model.text(),
+            'cellpose_model_type': self.cellpose_model_type.currentText(),
+            'cellpose_user_model': self.cellpose_user_model.text(),
+            'cellpose_diameter': self.cellpose_diameter.value(),
+            'cellpose_cellprob_threshold': self.cellpose_cellprob_threshold.text() if self.cellpose_cellprob_threshold.text() != '' else self.cellpose_cellprob_threshold.placeholderText(),
+            'cellpose_flow_threshold':  self.cellpose_flow_threshold.text() if self.cellpose_flow_threshold.text() != '' else self.cellpose_flow_threshold.placeholderText(),
             'output_user_suffix': self.output_user_suffix.text(),
             'channel_position': self.channel_position.value(),
             'projection_mode_bestZ': self.projection_mode_bestZ.isChecked(),
@@ -282,7 +335,11 @@ class Segmentation(QWidget):
         self.use_input_folder.setChecked(widgets_state['use_input_folder'])
         self.use_custom_folder.setChecked(widgets_state['use_custom_folder'])
         self.output_folder.setText(widgets_state['output_folder'])
-        self.selected_model.setText(widgets_state['selected_model'])
+        self.cellpose_model_type.setCurrentText(widgets_state['cellpose_model_type'])
+        self.cellpose_user_model.setText(widgets_state['cellpose_user_model'])
+        self.cellpose_diameter.setValue(widgets_state['cellpose_diameter'])
+        self.cellpose_cellprob_threshold.setText(widgets_state['cellpose_cellprob_threshold'])
+        self.cellpose_flow_threshold.setText(widgets_state['cellpose_flow_threshold'])
         self.output_user_suffix.setText(widgets_state['output_user_suffix'])
         self.channel_position.setValue(widgets_state['channel_position'])
         self.projection_mode_bestZ.setChecked(widgets_state['projection_mode_bestZ'])
@@ -303,32 +360,6 @@ class Segmentation(QWidget):
         Retrieve the input parameters
         Iterate over the image paths given performing f.main() function
         """
-        def check_inputs(image_paths, model_path, output_paths, output_basenames):
-            if len(image_paths) == 0:
-                self.logger.error('Image missing')
-                return False
-            for path in image_paths:
-                if not os.path.isfile(path):
-                    self.logger.error('Image not found: %s', path)
-                    return False
-            if model_path == '':
-                self.logger.error('Model missing')
-                self.selected_model.setFocus()
-                return False
-            if not os.path.isfile(model_path):
-                self.logger.error('Model not found: %s', model_path)
-                self.selected_model.setFocus()
-                return False
-            if self.output_folder.text() == '' and not self.use_input_folder.isChecked():
-                self.logger.error('Output folder missing')
-                self.output_folder.setFocus()
-                return False
-            output_files = [os.path.join(d, f) for d, f in zip(output_paths, output_basenames)]
-            duplicates = [x for x, y in zip(image_paths, output_files) if output_files.count(y) > 1]
-            if len(duplicates) > 0:
-                self.logger.error('More than one input file will output to the same file (output files will be overwritten).\nEither use input image folder as output folder or avoid processing images from different input folders.\nProblematic input files:\n%s', '\n'.join(duplicates[:4] + (['...'] if len(duplicates) > 4 else [])))
-                return False
-            return True
 
         channel_position = self.channel_position.value()
         projection_type = self.projection_type.currentText()
@@ -342,7 +373,11 @@ class Segmentation(QWidget):
             projection_zrange = None
 
         image_paths = self.image_list.get_file_list()
-        model_path = self.selected_model.text()
+        model_type = self.cellpose_model_type.currentText()
+        diameter = self.cellpose_diameter.value()
+        model_path = self.cellpose_user_model.text()
+        cellprob_threshold = float(self.cellpose_cellprob_threshold.text()) if self.cellpose_cellprob_threshold.text() != '' else float(self.cellpose_cellprob_threshold.placeholderText())
+        flow_threshold = float(self.cellpose_flow_threshold.text()) if self.cellpose_flow_threshold.text() != '' else float(self.cellpose_flow_threshold.placeholderText())
         user_suffix = self.output_user_suffix.text()
         output_basenames = [gf.splitext(os.path.basename(path))[0] + self.output_suffix + user_suffix for path in image_paths]
         if self.use_input_folder.isChecked():
@@ -350,7 +385,35 @@ class Segmentation(QWidget):
         else:
             output_paths = [self.output_folder.text() for path in image_paths]
 
-        if not check_inputs(image_paths, model_path, output_paths, output_basenames):
+        if len(image_paths) == 0:
+            self.logger.error('Image missing')
+            return
+        for path in image_paths:
+            if not os.path.isfile(path):
+                self.logger.error('Image not found: %s', path)
+                return
+        if model_type == "User trained model":
+            if model_path == '':
+                self.logger.error('Model missing')
+                self.cellpose_user_model.setFocus()
+                return
+            if not os.path.isfile(model_path):
+                self.logger.error('Model not found: %s', model_path)
+                self.cellpose_user_model.setFocus()
+                return
+        elif model_type not in ['cyto', 'cyto2', 'nuclei'] and diameter == 0:
+            self.logger.error('Diameter estimation using cellpose built-in model (i.e. diameter == 0) is only available for cyto, cyto2 and nuclei models')
+            self.cellpose_diameter.setFocus()
+            return
+
+        if self.output_folder.text() == '' and not self.use_input_folder.isChecked():
+            self.logger.error('Output folder missing')
+            self.output_folder.setFocus()
+            return
+        output_files = [os.path.join(d, f) for d, f in zip(output_paths, output_basenames)]
+        duplicates = [x for x, y in zip(image_paths, output_files) if output_files.count(y) > 1]
+        if len(duplicates) > 0:
+            self.logger.error('More than one input file will output to the same file (output files will be overwritten).\nEither use input image folder as output folder or avoid processing images from different input folders.\nProblematic input files:\n%s', '\n'.join(duplicates[:4] + (['...'] if len(duplicates) > 4 else [])))
             return
 
         # disable messagebox error handler
@@ -371,7 +434,6 @@ class Segmentation(QWidget):
         if self.use_gpu.isChecked():
             n_count = 1
             run_parallel = False
-
         QApplication.setOverrideCursor(QCursor(Qt.BusyCursor))
         for image_path, output_path, output_basename in zip(image_paths, output_paths, output_basenames):
             self.logger.info("Segmenting image %s", image_path)
@@ -379,7 +441,11 @@ class Segmentation(QWidget):
             QApplication.processEvents()
             arguments.append(
                 (image_path,
+                 model_type,
                  model_path,
+                 diameter,
+                 cellprob_threshold,
+                 flow_threshold,
                  output_path,
                  output_basename,
                  channel_position,